Layer SubstituteForBuilder on top of AutoSubstituteBuilder (#48)
This greatly simplifies some of the registration code as now there is no need to call .Configure(...) or .Configured() to continue registering other types. Since .Configured() was never released publicly, this is now removed, while .Configure(...) is obsoleted with a new API .ConfigureSubsitute introduced which returns the current SubstituteForBuilder.

The new succinct syntax is much more fluid, and doesn't break currently built tests.